+++ psycopg/psycopg/cursor.py
@@ -56,6 +56,8 @@ FATAL_ERROR = pq.ExecStatus.FATAL_ERROR
 SINGLE_TUPLE = pq.ExecStatus.SINGLE_TUPLE
 PIPELINE_ABORTED = pq.ExecStatus.PIPELINE_ABORTED
 
+ACTIVE = pq.TransactionStatus.ACTIVE
+
 
 class BaseCursor(Generic[ConnectionType, Row]):
     __slots__ = """
@@ -786,6 +788,14 @@ class Cursor(BaseCursor["Connection[Any]", Row]):
                     yield rec
                     first = False
         except e.Error as ex:
+            # try to get out of ACTIVE state. Just do a single attempt, which
+            # shoud work to recover from an error or query cancelled.
+            if self._pgconn.transaction_status == ACTIVE:
+                try:
+                    self._conn.wait(self._stream_fetchone_gen(first))
+                except Exception:
+                    pass
+
             raise ex.with_traceback(None)
 
     def fetchone(self) -> Optional[Row]: