+++ docs/news.rst
@@ -35,6 +35,7 @@ Psycopg 3.0.15 (unreleased)
 
 - Fix wrong escaping of unprintable chars in COPY (nonetheless correctly
   interpreted by PostgreSQL).
+- Restore the connection to usable state after an error in `~Cursor.stream()`.
 
 
 Current release